Fix history page. Add change tracking status and manager endpoints. Add user in order.

+++ src/main/java/com/winestore/domain/repository/product/WineRepository.java
@@ -8,8 +8,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
-
 @Repository
 public interface WineRepository extends CrudRepository<Wine, Long> {
     Page<Wine> findAll(Specification<Wine> specification, Pageable pageable);
@@ -18,5 +16,5 @@ public interface WineRepository extends CrudRepository<Wine, Long> {
         "from customer_review cr " +
         "where cr.wine_id = :wineId " +
         "and cr.confirm = true", nativeQuery = true)
-    BigDecimal countRating(Long wineId);
+    Integer countRating(Long wineId);
 }
